refactor(api): restructure db.rs and api_routes.rs into api_v1 module, err to common/http_errors
