+++ api-mock-server/src/api_v1/mod.rs
@@ -1,0 +1,2 @@
+pub mod db;
+pub mod routes;
